Add creation of aws ami images with pulumi
The AMI image creation is just a proof of concept at the moment.
Its not part of the normal aws or hetzner infrastructure cloud binary.

The pulumi binary packer implements the AMI image creation by creating
a wireguard ec2 instance without VPC and then open an ssh connection
to this instance and waiting for cloud-init to be finished. Then its
calling the AmiFromInstance method of ec2 pulumi resource and the AMI
image is created.

Major disadvantage at the moment is that the AMI image is part of the
pulumi state and there fore when this is destroyed the AMI image is
destroyed as well.