Implement the SSH connection as an actor model
The CreateImage function supports now an optional
connection provider to the virtual machine with an
actor model.

There is a poc implementation of an a SSH actor available `SSHConnector`.
The next steps would be to support file uploads multiple ssh commands
and so on to support full virtual machine provisioning with ssh.

And also keep this provisioning cloud provider independant as well.